Add loops and trim space.
Add loop when counting number of record. Only break the loop if the
number of record  is greater than 0
Trim empty space in MyFolder1
Loop when waiting for Download E-File windows appear. If error, close
and reload the window
Loop when waiting for SaveAs window appear
Loop when waiting for SaveAs window dis-appear